+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 1.0.0 - 20/01/2023
 
 * Initial release
+
+## 1.0.1 - 20/01/2023
+
+* Minor changes
+
+## 1.1.0 - 27/01/2023
+
+* Added support for icons in addition to labels
+* Now also supports icons with labels